+++ freamon/__init__.py
@@ -2,7 +2,7 @@
 Freamon: A package to make data science projects on tabular data easier.
 """
 
-__version__ = "0.2.9"
+__version__ = "0.3.0"
 
 # Import key components for convenient access
 from freamon.modeling.lightgbm import LightGBMModel